extract scoll to custom hook and refactor header to use functional compent and hooks
