+++ src/components/alphabetSlider.js
@@ -1,4 +1,5 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
+import { useScroll } from "../common/useScroll";
 import AlphabethSliderStyle from "./alphabetSlider.module.css";
 
 const chars = ["A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z"]
@@ -6,22 +7,7 @@ const chars = ["A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", 
 
 export default (props) => {
 
-  const [scrolled, setScrolled] = useState(false);
-  console.log(scrolled);
-  const scrollHandler = evt => {
-    if (evt.currentTarget.pageYOffset > 0) {
-      setScrolled(true);
-    } else {
-      setScrolled(false);
-    }
-  };
-
-  useEffect(() => {
-    window.addEventListener('scroll', scrollHandler);
-    return () => {
-      window.removeEventListener('scroll', scrollHandler);
-    }
-  })
+  const scrolled = useScroll();
 
   let style = AlphabethSliderStyle.container;
   style += (scrolled ? ' ' + AlphabethSliderStyle.scrolled : '');
